third_party/WebKit/LayoutTests/fast/scrolling/unscrollable-subpixel-overflow-mouse-drag-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
FAIL container.scrollTop should be 0. Was 0.5714285969734192.
Tests scroll behavior when an unscrollable scrollbar thumb is dragged by mouse. Passes if neither the thumb or the content changes.
